Rename emop.py command to emopcmd.py
This rename will avoid load conflicts with emop/ directory when generating docs and writing unit tests.

Update docs to contain documentation for emopcmd.py

Move logic that tried to guess pg_image_path from EmopJob to EmopPage model.  This is a more logical place for this code which will hopefully be removed later.

Use dedicated version of numpy and scipy to support DeNoise which seems to break with newer versions of NumPy and/or SciPy

Fix system tests to use Globus transfer.