Ocular support.  Ocular support requires Dashboard v2 API be used.
To support operating on works (books) instead of pages the controller was updated to modify behavior based on controller/operate_on config value.

Update Globus transfer logic to ensure only unique files are transfered and to auto-activate if one endpoint is expired or not activated.

Allow transfers to be disabled in emopcmd.py when using submit subcommand.

config.ini changes:
* Add controller/operate_on to control whether controller operates on pages or works (books)
* Set default scheduler/queue and scheduler/transfer_queue to use stakeholder,stakeholder-4g - reflect changes made on Brazos
* Remove modules section - not used
* Add ocular section with emission_engine = OPENCL to support Ocular